- finish end to end

+++ ratelimit-service/src/main/java/com/ayemin/ratelimit/service/CacheService.java
@@ -6,5 +6,5 @@ public interface CacheService {
      * @param key
      * @return
      */
-    int increase(String key);
+    long increase(String key);
 }
